Remove fetch calls from Graph board
